+++ dsmrreader/__init__.py
@@ -1,5 +1,5 @@
 from django.utils.version import get_version
 
-VERSION = (3, 3, 0, 'beta', 6)
+VERSION = (3, 3, 0, 'final', 0)
 
 __version__ = get_version(VERSION)