+++ docs/changelog.rst
@@ -9,6 +9,7 @@ v3.3.0 - 2020-xx-xx
 - [`#862 <https://github.com/dennissiemensma/dsmr-reader/issues/862>`_] Hogere backend process sleep toestaan
 - [`#864 <https://github.com/dennissiemensma/dsmr-reader/issues/864>`_] Requirements upgrade (2020-1)
 - [`#847 <https://github.com/dennissiemensma/dsmr-reader/issues/847>`_] Datalogger improvements
+- [`#869 <https://github.com/dennissiemensma/dsmr-reader/issues/869>`_] Sqlsequencereset bij deploy
 
 
 ----